Drop v3-stale image_dir surfaces and scope path layout to legacy module
`as_path()` / `from_path()` only make sense for the SDK v2 disk-cache
local-registry layout (`<root>/<image_name>/<tag>/`). v3 stores blobs
content-addressed under `<root>/blobs/sha256/...` and refs in SQLite,
so the v2-shaped per-image directory path is meaningless as a v3
storage location. Three public surfaces still leaked it to users:

- `ommx::artifact::get_image_dir(&ImageRef) -> PathBuf` (Rust)
- `ommx image-dir <name>` (CLI subcommand)
- `ommx.get_image_dir(image_name: str) -> PathBuf` (Python)

All three return a path that does not exist in v3 and never will, so
they're removed. The one legitimate v3-era use — the CLI's
"this artifact only exists in the legacy v2 cache, run
`ommx artifact import`" hint — switches to
`local_registry::legacy_local_registry_path`, which is the canonical
helper for "where would the v2 cache for this ref live?". The
deprecated `image_dir` alias goes away too.

The path-layout helpers themselves (`as_path` / `from_path`) move from
`ImageRef` into `local_registry::import::legacy` as crate-private
`image_ref_as_path` / `image_ref_from_path` functions, alongside the
existing `legacy_local_registry_path` they back. The `ImageRef`
module-level rustdoc, the v2 → v3 round-trip tests, and the
"no legacy dir left behind" assertion in
`local_registry/tests.rs` all switch to the new entry points.
