--- docs/api/_items/ommx.artifact.get_image_dir.rst
@@ -1,4 +1,0 @@
-get_image_dir
-=============
-
-.. pyo3-api-function:: ommx.artifact get_image_dir
