+++ python/ommx/ommx/artifact/__init__.py
@@ -6,7 +6,6 @@ from ommx._ommx_rust import (
     Artifact,
     ArtifactBuilder,
     Descriptor,
-    get_image_dir,
     get_images,
     get_local_registry_root,
     set_local_registry_root,
@@ -17,7 +16,6 @@ __all__ = [
     "Artifact",
     "ArtifactBuilder",
     "Descriptor",
-    "get_image_dir",
     "get_images",
     "get_local_registry_root",
     "set_local_registry_root",
